chore: implement decimal package

+++ cmd/api/request/add_product_request.go
@@ -5,8 +5,8 @@ import (
 )
 
 type AddProductRequest struct {
-	Name        string `json:"name" validate:"required,max=255"`
-	Description string `json:"description" validate:"required"`
-	Price       int64  `json:"price" validate:"required,gt=0,lte=9999999999"`
-	Quantity    int    `json:"quantity" validate:"required,gt=0,lte=9999"`
+	Name        string  `json:"name" validate:"required,max=255"`
+	Description string  `json:"description" validate:"required"`
+	Price       float64 `json:"price" validate:"required,gt=0,lte=9999999999"`
+	Quantity    int     `json:"quantity" validate:"required,gt=0,lte=9999"`
 }
